requirements.txt
omgifol
numpy
scikit-image